lambdas/jira_lambda.py
import json

def lambda_handler(event, context):
  try:
    if 'body' in event:
      post_data = json.loads(event['body'])
      result = process_jira_request(post_data)
    else:
      return {
        'statusCode': 400,
        'body': 'Bad Request',
      }

    return {
      'statusCode': 200,
      'body': json.dumps(result),
    }

  except Exception as e:
    return {
      'statusCode': 500,
      'body': json.dumps(str(e)),
    }

def process_jira_request(data):
  # Implement Jira-specific functionality using the provided data
  # Example: Create an issue in Jira based on the data
  # ...
  print(data)
